Two generate tasks in :test-cases-internal
