+++ .gitignore
@@ -47,3 +47,6 @@ generated/
 
 # Blueprint theme
 __init__.pyc
+
+verification-api/src/main/conjure/generated-client
+verification-api/src/main/conjure/generated-server
