+++ test-cases-internal/build.gradle
@@ -2,7 +2,31 @@ apply plugin: 'com.palantir.conjure'
 apply plugin: 'java-library'
 
 dependencies {
-    api 'com.palantir.conjure:conjure-core'
-    api project("${project.name}-objects")
-    api 'com.palantir.conjure.java.runtime:conjure-java-jackson-serialization'
+    compile 'com.fasterxml.jackson.core:jackson-databind'
+    compile 'com.fasterxml.jackson.dataformat:jackson-dataformat-yaml'
+    compile 'com.palantir.conjure:conjure-core'
+    compile 'com.palantir.conjure.java.runtime:conjure-java-jackson-serialization'
+    compile 'com.google.guava:guava'
+    compile project("${project.name}-objects")
+}
+
+task generateClientServices(type: JavaExec) {
+    main = "com.palantir.conjure.verification.GenerateClientServices"
+    classpath = sourceSets.main.runtimeClasspath
+    args rootProject.file('master-test-cases.yml'), '../verification-api/src/main/conjure/generated-client'
+    inputs.file rootProject.file('master-test-cases.yml')
+    outputs.dir 'src/main/conjure/generated-client'
+}
+
+task generateServerServices(type: JavaExec) {
+    main = "com.palantir.conjure.verification.GenerateServerServices"
+    classpath = sourceSets.main.runtimeClasspath
+    args rootProject.file('master-test-cases.yml'), '../verification-api/src/main/conjure/generated-server'
+    inputs.file rootProject.file('master-test-cases.yml')
+    outputs.dir 'src/main/conjure/generated-server'
+}
+
+task generate {
+    dependsOn generateClientServices
+    dependsOn generateServerServices
 }
