+++ test-cases-internal/src/main/java/com/palantir/conjure/verification/GenerateServerServices.java
@@ -1,47 +1,27 @@
 /*
  * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
  */
 
-package com.palantir.conjure.verification.server;
+package com.palantir.conjure.verification;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.io.MoreFiles;
 import com.google.common.io.RecursiveDeleteOption;
-import com.palantir.conjure.verification.BodyTests;
-import com.palantir.conjure.verification.MasterTestCases;
-import com.palantir.conjure.verification.ResolveLocalReferencesConjureTypeVisitor;
-import com.palantir.conjure.verification.SingleHeaderParamTests;
-import com.palantir.conjure.verification.SinglePathParamTests;
-import com.palantir.conjure.verification.SingleQueryParamTests;
-import com.palantir.conjure.verification.TestCasesUtils;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.List;
 import java.util.Map;
 
-public final class GenerateVerificationServerConjureDefinitions {
+public final class GenerateServerServices {
 
-    private GenerateVerificationServerConjureDefinitions() {}
+    private GenerateServerServices() {}
 
     public static void main(String... args) throws IOException {
-        Preconditions.checkArgument(args.length == 2, "Usage: <master-test-cases.yml> <conjure dir>");
+        Preconditions.checkArgument(args.length == 2, "Usage: <master-test-cases.yml> <gitignored output dir>");
         File file = new File(args[0]);
-        File conjureDir = new File(args[1]);
-        File outputDir = new File(conjureDir, "generated");
+        File outputDir = new File(args[1]);
 
         MasterTestCases testCases = TestCasesUtils.parseTestCases(file);
