--- verification-server-api/src/main/conjure/.gitignore
@@ -1,1 +1,0 @@
-/generated
